devices: virtio-fs: rework setattr Data::Handle enum
This previously held a raw descriptor (which is just an integer and does
not have an associated lifetime) and an Arc reference to its
corresponding HandleData in an attempt to ensure the HandleData did not
get destroyed while the RawDescriptor was alive. However, newer clippy
versions warn about the unused Arc<HandleDataField>, and this loose
association between the HandleData instance and the descriptor is prone
to being removed by accident since there is no actual lifetime.

To avoid this, keep only a MutexGuard<File> that refers to the file
inside the HandleData, and use a let binding to ensure that the Mutex
itself is alive for the remainder of the function (enforced with actual
Rust lifetimes rather than hopes and dreams).

BUG=b:344974550
TEST=tools/clippy

Change-Id: I4b1f66a2166422a87452fae597c3f08cbdc7dbe5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5623301
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>