rutabaga_gfx: fix MinigbmBuffer dead_code warning
Avoids a clippy warning about the unused second field by renaming it to
have an underscore prefix.

It seems like the MinigbmDevice is only used to ensure that the
gbm_device is not freed while buffers that were allocated from it are
still live, so it is OK that it's unused.

BUG=b:344974550
TEST=tools/clippy

Change-Id: I53aee94d4b366cb25cbc3a360131345c05670aa8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5626405
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>