arch: remove host OS checks for guest_suspended_cvar
The VirtualPmc setup code does not require any OS-specific
functionality, only a portable Condvar, so we can clean up some
unnecessary conditional compilation in the arch crates.

The force_s2idle feature is not actually wired up on Windows yet, but
that should also be possible to enable in an OS-agnostic way if desired.

BUG=b:194391015
TEST=tools/dev_container tools/presubmit

Change-Id: Iceee1068c20d64539063387b1073ff71c5d05fec
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5671243
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>