ext2: Fix permission of the root directory
When a source directory path is given to the ext2 logic, its metadata
such as mode should be copied to the disk.

Before this patch, some unit tests that dump files in the created file
system with `debugfs` failed to clean up these files after completion
because the top directory permission is not set properly.

BUG=b:351994509
TEST=CQ

Change-Id: I9315d8cae45a738e8abcda1a4313a89d5fa7bd16
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5687140
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>