devices: pci: remove unreachable 64-bit BAR check
end_addr is a u64, so its value cannot be greater than the max u64
value. The compiler doesn't warn about this when using the legacy
u64::max_value() function, but it becomes a warning when this is
replaced with the u64::MAX constant.

BUG=None
TEST=tools/dev_container tools/presumbit

Change-Id: I05ed36027718d62782ea2f2c597fe391fe9f0fae
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5690373
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>