hypervisor: x86_64: clarify Segment limit units
The limit value in a kvm_segment is intended to be in bytes, meaning it
should already include the scale factor from the G (granularity) field.
This is basically undocumented, but this is based on observation of the
values assigned to kvm_segment limit fields inside the kernel being
larger than 20 bits.

Rename the field in the generic Segment struct to "limit_bytes" to make
its units clear and match the KVM behavior.

I could not find documentation for either WHPX or HAXM to determine
whether their limit fields should be in units of bytes or not, so those
hypervisors currently perform the conversion to maintain the previous
behavior of using the raw limit field as encoded in the GDT entry.

This mostly does not matter for 64-bit guests, since segment limit
fields are ignored in long mode, but it does matter in 32-bit protected
mode.

BUG=b:354053941

Change-Id: I1e464d3e1b8556184c732adea034b326e5ed6a72
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5727075
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>