bindgen: Update KVM_PVIOMMU_SET_CONFIG ioctl num
Update the ioctl number to a well-formed value, following the similar
fix to the experimental Android kernel. Obviously, this is a breaking
change rendering new (resp. old) versions of that kernel incompatible
with version of crosvm missing (resp. including) this commit.

However, this is fine because both the kernel and this support for
pKVM-VFIO are experimental, as stated in fdb527c202cd ("vfio: Add
support for pKVM-VFIO pvIOMMUs").

TEST=Commit-Queue
Bug=b:278039964
BUG=b:306110430

Change-Id: Iefdebfb3df41aa647fac43eab50e379ca7821817
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5739077
Auto-Submit: Pierre-Clément Tosi <ptosi@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Pierre-Clément Tosi <ptosi@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>