crosvm: add option to retry on vhost-user connect failure
When enabled and the first attempt fails, prints out a warning and then
silently retries for the full duration.

When starting crosvm with some vhost-user devices, there isn't an
explicit signal from the vhost-user backend process that it is ready and
listening. Often people will just start the backend first, then the
frontend and hope it works, but there is technically a race condition.
To be reliable, the layer above would need to probe the socket path
somehow. It would be best to take care of that logic as part of the
vhost-user frontend.

Instead of polling, we could consider something like inotify on Linux,
but it is more complex and might not be entirely sufficient by itself
because sometimes an old socket file is left around with no listener
(causing a ConnectionRefused error).

BUG=b:340510697
TESTED=manually started the frontend, confirmed the warning log, then started the backend

Change-Id: Id446a690cc64a4a4e100b4cd7df140ddcf17e7d0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5754390
Reviewed-by: Jason Macnak <natsu@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>