devices: vmwdt: fix potential timer deadlock
Use the `mark_waited` method (non-blocking) instead of `wait`
(blocking).

It is technically possible for a vCPU thread to disable a timer after it
has fired but before the worker successfully calls `wait`. In that case,
the worker would get stuck and be unable process timer events for other
vCPUs and also would be unable to shutdown. I haven't seen this happen
in practice.

Change-Id: I49e18ba525392a08f5dc1e22181876cf6a013173
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5825223
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Sebastian Ene <sebastianene@google.com>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>