Make pvclock suspend logic happen after vcpu kick
In b/345278018, negative time deltas were observed by the guest kernel
as we were overestimating the suspend duration, which in turn was
because we started the measurement before we kicked to suspend VCPUs.

Ensure that suspend time measurements starts after VCPU suspends and
ends before VCPU resumes so that our measurement is fully conservative
and does not result in a negative delta.

BUG=b:345278018
TEST=tast run arc.Suspend.s10c2 on kukui (aarch64)

Change-Id: Ieb87a31ea1bce94e2d667032cdaea0710e625fe3
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5823030
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: Tatsuyuki Ishi <ishitatsuyuki@google.com>