x86_64: do not write to boot_params::cmdline_size
The cmdline_size field in the Linux x86 boot protocol is meant to be
read only; the boot loader (crosvm in this case) should read it to
determine the maximum command line size, but it should not be
overwritten. The kernel determines the command line length via the NUL
terminator byte, not by checking the cmdline_size field.

<https://www.kernel.org/doc/html/latest/arch/x86/boot.html>

No functional change (Linux kernels do not read this field at runtime).

BUG=b:362168475

Change-Id: I8f4932b2c9a8e66948f56f12d5b40f58b0656763
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5838024
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>