pmem-ext2: Create ext2 filesystem after minijail fork
Running mkfs can take a long time because it scans the given directory
entirely.
Running it in the main process blocks the initialization of other devices.
Also, since minijail's fork does not allow threads to exist in the same
process, we can't spawn a thread in the main process.

So, with this CL, we fork a new jailed process for creating a filesystem
on memory as well as the existing pmem device process.
This allows ext2 to be created in parallel with the VM boot process.
Also, we can restrict pmem's device process and the new filesystem
creation process with different seccomp policy.

You can find sequence diagram of the processes in go/virtio-pmem-ext2.

BUG=b:329359333
TEST=run VM on workstation

Change-Id: I60c12caddb22fc59389c317dea7e1102a4772e01
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5644847
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>