arch: add optional param to load_image_high() to filter memory regions
Add an optional filter function param to load_image_high() for extra
filtering of guest memory regions when searching for a suitable region
to load the image to. This is useful in cases when we want to prevent
loading the image to some of the regions located between
min_guest_addr and max_guest_addr, as we already have those regions
reserved for something else.

This will be used to ensure that initrd doesn't overlap with pvmfw
on x86.

BUG=b:354676864
TEST=run a VM with initrd

Change-Id: I9dc4f3e1f9cfe8eb02cb6f7e6e3564924b56b16e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5872814
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>