devices: vhost-user frontend: shorten thread name
The worker thread used to be named "vhost_user_virtio_{type}", which is
too long to fix in the 15 characters available for thread names on
Linux. Use the device's debug_label() instead, which follows the pattern
"vu-{type}", so the threads for different devices can be distinguished.

Change-Id: I8cadf54518724c5a764a0b28b0bc4144cd938d1e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5906571
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>