crosvm: refactor tube plumbing in vm init
This change defines a new enum that captures most of the types of tubes
connecting the devices and the control threads. During device creation,
all the tubes are gathered into one big list and then, in `run_control`,
they are sorted out into separate lists based on the implementation
details of the control threads.

Change-Id: I11de36543d6a4b07acd88b20f96e9690c843d78b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5923038
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>