x86_64: move PCI memory layout logic out of global
No behavior changes intended.

A new `ArchMemoryLayout` type is introduced to support arch specific
memory layout details. It is created after `VmComponents`, but before
`GuestMemory`. On x86_64 it is used to carry the information that was
previously stored in a global variable. On aarch64 and riscv64, it is
empty for now.

BUG=b:361390145

Change-Id: I1e3b665a1bccf0de04a4a4576e6fb014803e835b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5938328
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>