crosvm_control: make validate_socket_path() unsafe
The validate_socket_path() is not a safe function. Checking socket_path
not null is not enough for stating CStr::from_ptr(socket_path) is safe.
The caller also needs to ensure the socket_path points to a
nul-terminated memory and the string size should not exceed isize::MAX.

BUG=b:353609634
TEST=cargo build

Change-Id: Ia32552e5988365c2b2f3f7e522f7e0a979b63809
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5955891
Commit-Queue: Yuan Yao <yuanyaogoog@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>