virtio-pmem: switch allocation to bottom up
When using a Linux guest kernel with kernel commit
ea72ce5da22806d5713f3ffb39a6d5ae73841f93 ("x86/kaslr: Expose and use
the end of the physical memory address space"), the pmem region can't be
mapped, because it is beyond PHYSMEM_END after the kaslr code has
adjusted it. This breaks reading from pmem on affected systems, since
the guest can't map the memory.

As a short-term workaround, we can change the allocation strategy so
that the pmem region gets allocated from the lowest available address
instead of the highest. This can technically result in the pmem region
being allocated inside the PCI host controller window, but this works in
practice for the cases we care about.

BUG=b:375506171
TEST=Boot Crostini with rootfs on /dev/pmem0

Change-Id: I8f301597f7d4847aa4da87050f1e7d5f22ae39e9
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5980378
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>