x86_64: bzImage: load extra data appended to kernel
The bzImage kernel loader would previously only load the amount of data
specified in the header via the syssize field. In a normal Linux kernel,
this matches the size of the file exactly, but it will be convenient in
some cases to load a kernel with an additional footer appended, so the
bzImage code is extended to load everything in the file after the
setup_sects offset into guest memory.

BUG=b:372755978
TEST=load bzImage with extra text appended and verify with gdb

Change-Id: I102aead8d89b68b6932ac1fcc30ccc5d5e769d1b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5983885
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>