disk: composite: fix secondary GPT header location
The secondary GPT header should be at the end of the image, but its
offset was chosen before the disk size was aligned to 64k. This caused
Linux to print warnings at boot, like

    GPT:Primary header thinks Alt. header is not at the end of the disk.
    GPT:105736 != 105855
    GPT:Alternate GPT header not at the end of the disk.
    GPT:105736 != 105855
    GPT: Use GNU Parted to correct GPT errors.

It is likely there was no practical issue besides the log noise.

As part of the fix, we no longer align the disk size to 64k. It is
unclear why we aligned to 64k to begin with and maintaining that
alignment made the appropriate fix more complicated.

Change-Id: I2f41a2033f14d34e2fd3dd3dd3afe2da80266196
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6036619
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>